backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
Markdown==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
